Remove custom version of dockervoyager.pp.
We don't want to remove the general custom_zulip_files toolchain, but
we've now integrated a correct version of dockervoyager.pp into
zulip.git, so we don't need to include this patched version (which is
also not compatible with Puppet 4).

This is possible thanks to the zulip.git branch ending in
https://github.com/zulip/zulip/commit/94982605162694f9c38d85b6674d2cbefc3bf570
that was recently merged.