Always generate secrets and do so more safely
Firstly, we no long persist `zulip-secrets.conf` to the data directory.
Since it's generated on every boot, we just keep it in `/etc` where
Zulip wants it.

Secondly, we no longer replace the old secret's assignment in place,
which had an assumption which prohibited at least ~ from being in the
passphrase. Instead, we delete the assignment, if it's there, and append
it to the config anew.