Add helm chart
A helm chart is added for the installation of Zulip into Kubernetes
clusters. The chart is based on a version of PR #287, but contains
significant changes.

The chart depends on Bitnami helm charts for the installation of
Memcached, Postgresql, Rabbitmq and Redis.

The chart includes the possibility to add `postSetup` scripts which
are mounted to `/data/post-setup.d` and executed by the docker-zulip
Docker container.

More information can be found in the chart's README in
kubernetes/chart/zulip/README.md